requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
Flask==3.1.0
flask-ldap3-login==1.0.2
Flask-WTF==1.2.2
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
ldap3==2.9.1
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.1
pyasn1==0.6.1
pycparser==2.22
PyYAML==6.0.2
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.18.3
